LayoutTests/fast/forms/date/date-editable-components/date-readonly-picker-does-not-open-on-space-keypress-expected.txt

Tests that a readonly date input's calendar view cannot be opened using the space key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS showingPicker is false
Focusing readonly input using tab key and then pressing space key.
PASS showingPicker is false
PASS document.getElementById('input').value is "1970-01-01"
Removing readonly attribute and pressing space key.
PASS showingPicker is true
PASS successfullyParsed is true

TEST COMPLETE
